+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.10] - 2026-04-20
+
+### Docs
+- Update README.md
+
+### Other
+- Update redeploy/cli.py
+- Update redeploy/spec_loader.py
+- Update redeploy/tests/test_cli.py
+- Update redeploy/tests/test_spec_loader.py
+
 ## [0.2.9] - 2026-04-20
 
 ### Docs